Последний для проверки коммит в k2_dz4 ветку ДЗ-4 Spring Boot skypro-spring-k2-dz4-nikrink

+++ src/main/java/pro/sky/skyprospringk2dz4nikrink/CalcService.java
@@ -1,11 +1,13 @@
 package pro.sky.skyprospringk2dz4nikrink;
 
-import org.springframework.web.bind.annotation.RequestParam;
-
 public interface CalcService {
     String calculator();
-    String plus (Integer num1, Integer num2);
-    String minus (Integer num1, Integer num2);
-    String multiply (Integer num1, Integer num2);
-    String divide (Double num1, Double num2);
+
+    String plus(Integer num1, Integer num2);
+
+    String minus(Integer num1, Integer num2);
+
+    String multiply(Integer num1, Integer num2);
+
+    String divide(Integer num1, Integer num2);
 }
